// typescript/user-interface/packages/ui-state/src/ts/app/state/waveform/types.ts
export interface StationVisibilityChanges {
  /**
   * Whether this station is visible.
   */
  visibility: boolean;

  /**
   * The name of this station. Used to look up the station from the results received from the server,
   * and assumed to be unique.
   */
  stationName: string;

  /**
   * If not set, default to true.
   */
  hiddenChannels?: string[];

  /**
   * Whether the station is expanded to show all visible channels. Defaults to false.
   */
  isStationExpanded?: boolean;
}
// ...
export type StationVisibilityChangesDictionary = Record<string, StationVisibilityChanges>;

// State of loading waveforms
export interface WaveformLoadingState {
  isLoading: boolean;
  total: number;
  completed: number;
  percent: number;
  description: string;
}

/**
 * Waveform Redux State
 */
export interface WaveformState {
  /** maps the names of channels and signal detection ids (which have derived channels) to the filter that is applied */
  channelFilters: ChannelFilterRecord;
  /** maskVisibility */
  maskVisibility: Record<string, boolean>;
  /** tracks the changes to the default visibility in the waveform display */
  stationsVisibility: StationVisibilityChangesDictionary;
  /** tracks the station names in the currently open station group */
  currentStationGroupStationNames: string[];
  /** tracks how many requests are in flight, and how many are complete, for loading waveform data */
  loadingState: WaveformLoadingState;
  // whether or not to show the time uncertainty bars in the waveform display.
  shouldShowTimeUncertainty: boolean;
  // whether or not to show predicted phases in the waveform display.
  shouldShowPredictedPhases: boolean;
  // The amount that is currently in on the screen due to zooming.
  zoomInterval: CommonTypes.TimeRange;
  // The amount of time that can be viewed without loading more data.
  // because the user may load more waveform
  // data than the currently opened time interval via panning.
  viewableInterval: CommonTypes.TimeRange;
  // The offsets when not aligned by time
  minimumOffset: number;
  maximumOffset: number;
  baseStationTime: number;
  displayedSignalDetectionConfiguration: Record<
    WaveformDisplayedSignalDetectionConfigurationEnum,
    boolean
  >;
  viewportVisibleStations: string[];
}
// ...
export enum WaveformDisplayedSignalDetectionConfigurationEnum {
  signalDetectionBeforeInterval = 'signalDetectionBeforeInterval',
  signalDetectionAfterInterval = 'signalDetectionAfterInterval',
  signalDetectionAssociatedToOpenEvent = 'signalDetectionAssociatedToOpenEvent',
  signalDetectionAssociatedToCompletedEvent = 'signalDetectionAssociatedToCompletedEvent',
  signalDetectionAssociatedToOtherEvent = 'signalDetectionAssociatedToOtherEvent',
  signalDetectionConflicts = 'signalDetectionConflicts',
  signalDetectionDeleted = 'signalDetectionDeleted',
  signalDetectionUnassociated = 'signalDetectionUnassociated'
}